ARM64: Decrease the sizes of some values
These values can be changed to smaller types without causing problems.

I've done this in a separate commit because my dev setup doesn't use
ARM. In case something ARM-specific breaks in CI, I'd like to be able to
keep those changes in their own commit so I can debug more easily.

+++ src/vmm/src/arch/aarch64/cache_info.rs
@@ -37,7 +37,7 @@ pub(crate) struct CacheEntry {
     pub level: u8,
     // Type of cache: Unified, Data, Instruction.
     pub type_: CacheType,
-    pub size_: Option<usize>,
+    pub size_: Option<u32>,
     pub number_of_sets: Option<u16>,
     pub line_size: Option<u16>,
     // How many CPUS share this cache.
@@ -211,12 +211,12 @@ fn readln_special<T: AsRef<Path>>(file_path: &T) -> Result<String, CacheInfoErro
     Ok(line.trim_end().to_string())
 }
 
-fn to_bytes(cache_size_pretty: &mut String) -> Result<usize, CacheInfoError> {
+fn to_bytes(cache_size_pretty: &mut String) -> Result<u32, CacheInfoError> {
     match cache_size_pretty.pop() {
-        Some('K') => Ok(cache_size_pretty.parse::<usize>().map_err(|err| {
+        Some('K') => Ok(cache_size_pretty.parse::<u32>().map_err(|err| {
             CacheInfoError::InvalidCacheAttr("size".to_string(), err.to_string())
         })? * 1024),
-        Some('M') => Ok(cache_size_pretty.parse::<usize>().map_err(|err| {
+        Some('M') => Ok(cache_size_pretty.parse::<u32>().map_err(|err| {
             CacheInfoError::InvalidCacheAttr("size".to_string(), err.to_string())
         })? * 1024
             * 1024),
